refactor(cli): A1 phase 2 — single-source-of-truth slash command registry
Every slash command now lives in one of eleven category modules under
``libs/cli/bog_agents_cli/commands/``. The legacy
``BogAgentsApp._COMMAND_HANDLER_NAMES`` map and the ~575-line literal
``SLASH_COMMAND_SPECS`` tuple in ``command_registry`` are gone. Adding
or renaming a slash command means editing exactly one module — drift
between dispatch, autocomplete, and help is mechanically impossible.

What changed:
  - Extract SlashCommandSpec into bog_agents_cli/_spec.py so command
    modules can import it without an import cycle through
    command_registry.
  - command_registry.SLASH_COMMAND_SPECS is now derived from
    ``commands.COMMANDS`` (which itself is built by walking
    ``commands/`` at import time).
  - Remove BogAgentsApp._COMMAND_HANDLER_NAMES; _resolve_command_handler
    consults the registry's COMMAND_HANDLER_MAP exclusively.
  - 11 new command modules (agent, analysis, config, enterprise,
    general, git, info, multimodal, quality, ui, web) plus the existing
    session module — 75 commands total, with /q, /cost, /context aliases
    flowing through SlashCommandSpec.aliases.
  - Order: featured commands first (matching /help showcase), then
    alphabetical for the long tail. _registry.py owns the policy.
  - Autocomplete tie-break by name length so /help wins over /health
    when the user types /he.
  - Tests rewritten: test_commands_registry covers uniqueness,
    handler-method existence, alias resolution; test_app's existing
    coverage check now consults COMMAND_HANDLER_MAP.

Regression results:
  SDK:    1344 passed
  CLI:    2689 passed (full suite minus pre-existing flaky test_cmd_undo)
  Daemon:   95 passed
  Lint:   ruff clean across all three packages
