+++ libs/cli/bog_agents_cli/command_registry.py
@@ -3,576 +3,23 @@
 This module is intentionally lightweight so it can be imported from startup
 paths like help rendering, autocomplete wiring, and command-palette helpers
 without pulling in the SDK or Textual runtime.
+
+Spec data is sourced from the modular ``bog_agents_cli/commands/`` package.
+Adding a new slash command means adding one module under ``commands/``;
+this file exposes the search / palette / autocomplete helpers that build
+on top of that data plus extension-contributed commands loaded at runtime.
 """
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 from difflib import get_close_matches
 from pathlib import Path
 
+from bog_agents_cli._spec import SlashCommandSpec
+from bog_agents_cli.commands import COMMANDS as _COMMANDS
 
-@dataclass(frozen=True, slots=True)
-class SlashCommandSpec:
-    """Metadata for one slash command."""
+SLASH_COMMAND_SPECS: tuple[SlashCommandSpec, ...] = tuple(c.spec for c in _COMMANDS)
 
-    name: str
-    description: str
-    hidden_keywords: str = ""
-    category: str = "general"
-    shortcut: str = ""
-    aliases: tuple[str, ...] = ()
-    available: bool = False
-
-
-SLASH_COMMAND_SPECS: tuple[SlashCommandSpec, ...] = (
-    SlashCommandSpec(
-        "/help",
-        "Show slash command help and search by keyword",
-        "commands reference",
-        "general",
-        "?",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/agent",
-        "Manage parallel agent threads and worktrees (list/spawn/panel/switch/stop)",
-        "thread multi parallel worktree watch dashboard live status panel",
-        "agent",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/api",
-        "Send API requests and test endpoints",
-        "http rest curl",
-        "web",
-    ),
-    SlashCommandSpec(
-        "/audit",
-        "Audit dependencies for vulnerabilities",
-        "security deps",
-        "quality",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/background",
-        "Manage background agent tasks (run/list/cancel/status)",
-        "bg task async",
-        "agent",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/branch",
-        "Manage git branches for local workflows",
-        "git checkout switch create",
-        "git",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/changelog",
-        "Open the project changelog in your browser",
-        "release notes",
-        "info",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/clear",
-        "Clear chat history and start a fresh thread",
-        "reset new conversation",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/commands",
-        "Browse available slash commands and quick descriptions",
-        "help reference discover",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/compact",
-        "Summarize conversation to reduce context usage",
-        "retain keep drop summarize",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/dashboard",
-        "Show the multi-agent dashboard with status and costs",
-        "agents monitor panel",
-        "agent",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/diff",
-        "Show pending file changes as a diff",
-        "changes git",
-        "git",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/docs",
-        "Open documentation and project guides",
-        "readme api",
-        "info",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/doctor",
-        "Run health check diagnostics for the local CLI environment",
-        "check status",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/effort",
-        "Set effort level (low/medium/high/max)",
-        "quality speed",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/extensions",
-        "Manage extensions and extensibility packages",
-        "plugins marketplace",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/feedback",
-        "Open the issue tracker to report a bug or request a feature",
-        "bug issue request",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/harbor",
-        "Harbor benchmark evaluation — run tasks, view results, check evaluation status",
-        "benchmark eval terminal-bench evaluation trajectory langsmith",
-        "analysis",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/health",
-        "Codebase health score and analysis",
-        "quality complexity coverage",
-        "analysis",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/image",
-        "Analyze images or paste from clipboard",
-        "screenshot multimodal",
-        "multimodal",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/init",
-        "Generate `AGENTS.md` for the current repository",
-        "setup agents onboard",
-        "info",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/infra",
-        "Generate infrastructure code (Docker/K8s/Terraform)",
-        "devops deploy",
-        "analysis",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/keybindings",
-        "Show current keybindings or the config file path",
-        "keys shortcuts",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/logs",
-        "Show the log file path and recent warnings or errors",
-        "debug trace errors",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/mcp",
-        "MCP server marketplace — browse, install, and manage servers (jira, terraform, github…)",
-        "servers tools install registry catalog marketplace plugin integration",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/migrate",
-        "Plan technology migration",
-        "upgrade convert",
-        "analysis",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/model",
-        "Switch models or manage the default model",
-        "provider swap ollama",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/model-route",
-        "Configure automatic model routing",
-        "auto cost optimize",
-        "config",
-    ),
-    SlashCommandSpec(
-        "/onboard",
-        "Start an interactive codebase onboarding guide",
-        "tour walkthrough new",
-        "info",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/permissions",
-        "Show approval mode and shell permission settings",
-        "safety approvals shell trust",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/pipeline",
-        "Run a saved pipeline (chained prompts, skills, slash commands)",
-        "workflow chain schedule cron automate steps yaml",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/plan",
-        "Toggle read-only plan mode",
-        "readonly architect",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/plugin",
-        "Manage plugins and extensions — includes Claude Code skill import and MCP sync",
-        "marketplace skills extensions claude-code mcp sync import compatible",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/prompt",
-        "Browse and run saved prompts with variable substitution",
-        "template library saved custom variable",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/pr",
-        "Pull request management (create/list/review)",
-        "github merge",
-        "git",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/preview",
-        "Start or stop local dev server preview",
-        "serve browser",
-        "web",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/profile",
-        "Switch configuration profile",
-        "config preset",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/quit",
-        "Exit the app",
-        "close leave",
-        "general",
-        aliases=("/q",),
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/recommend",
-        "Run AI-powered code review and recommendation flows",
-        "review audit advise persona focus",
-        "quality",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/record",
-        "Start or stop recording session for replay",
-        "capture",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/repomap",
-        "Show or refresh the semantic repository map (classes, functions, imports)",
-        "index codebase symbols architecture structure repo-map",
-        "analysis",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/reload",
-        "Reload config from environment variables and `.env`",
-        "refresh",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/remember",
-        "Update memory and skills from the current conversation",
-        "memory skills capture",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/remote",
-        "Submit a task for remote or cloud execution",
-        "cloud",
-        "web",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/rewind",
-        "Browse checkpoints and fork a thread from an earlier snapshot",
-        "checkpoint recover restore history",
-        "info",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/resume",
-        "Resume a saved thread by id, tag, project, or browse history",
-        "continue switch history recover",
-        "info",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/replay",
-        "Replay agent actions for debugging",
-        "debug trace",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/resolve",
-        "AI-assisted merge conflict resolution",
-        "conflict merge",
-        "git",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/review",
-        "Ask the agent for a structured code review",
-        "lint check staged files commit",
-        "quality",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/session",
-        "Show or update session label, tags, project, summary, and exports",
-        "name duration info metadata",
-        "info",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/settings",
-        "Configure providers, models, and fallbacks",
-        "config preferences setup",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/silent",
-        "Quiet mode: tool calls show as one-liners (full details still in log)",
-        "quiet collapse mute hide",
-        "ui",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/verbose",
-        "Verbose mode: tool calls show as expandable widgets (default)",
-        "loud expand show",
-        "ui",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/skills",
-        "Show loaded skills and their search paths",
-        "abilities memory",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/teach",
-        "Start teaching mode to learn a workflow",
-        "learn skill",
-        "general",
-    ),
-    SlashCommandSpec(
-        "/team",
-        "Team shared config (members, context, prompts) and multi-agent coordination",
-        "enterprise org swarm coordination invite members shared context prompts config",
-        "enterprise",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/think",
-        "Enable extended thinking / deep reasoning on the next query",
-        "reasoning cot chain-of-thought extended-thinking budget tokens anthropic gemini",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/rules",
-        "Manage project rules (.bog-agents/rules/) — auto-inject contextual guidelines",
-        "mdc cursor-rules guidelines standards glob always inject frontmatter",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/search",
-        "Hybrid codebase search — ripgrep exact + fuzzy filename + semantic",
-        "ripgrep rg find grep vector embeddings semantic hybrid",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/compress",
-        "Intelligent context compression — auto-compact with ratio report and progress",
-        "compact context window tokens summarize reduce packing",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/jobs",
-        "Manage autonomous background jobs — list, attach, cancel, and monitor long-running tasks",
-        "background async detached autonomous worktree notification webhook",
-        "agent",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/langsmith",
-        "LangSmith integration — browse traces, runs, evals, datasets, feedback, and OTEL setup",
-        "tracing observability langchain eval runs traces feedback otel opentelemetry langchain-hub",
-        "analysis",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/workspace",
-        "Multi-repository context — define repos in .bog-agents/workspace.toml",
-        "multi-repo cross-repo microservices monorepo symbol resolution",
-        "enterprise",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/worktrees",
-        "Manage parallel git worktrees — spawn agents in isolated branches",
-        "parallel agents git worktree branch isolation conflict merge",
-        "enterprise",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/test",
-        "Run tests with coverage and generate test skeletons",
-        "coverage pytest",
-        "quality",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/threads",
-        "Browse and resume previous threads",
-        "continue history sessions",
-        "info",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/tokens",
-        "Show current token usage and context breakdown",
-        "cost context window budget spend",
-        "info",
-        aliases=("/cost", "/context"),
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/trace",
-        "Open the current thread in LangSmith",
-        "langsmith observability",
-        "info",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/undo",
-        "Inspect or restore tracked file changes with git",
-        "revert rollback",
-        "git",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/vars",
-        "Manage secrets and variables (API keys, URLs, tokens)",
-        "secrets env config keys tokens credentials",
-        "config",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/version",
-        "Show CLI and SDK versions",
-        "build release",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/worktree",
-        "Manage git worktrees for isolated work",
-        "isolate parallel",
-        "git",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/checkpoint",
-        "Save, load, and list named session checkpoints for easy resume",
-        "save restore resume snapshot session history",
-        "info",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/benchmark",
-        "Run evaluation suites and view recent benchmark results",
-        "eval score perf suite tasks trajectory",
-        "analysis",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/build",
-        "Interactive wizard — create skills, prompts, and pipelines step by step",
-        "wizard create new template scaffold variablize builder",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/explain",
-        "Deep-dive explanation of any symbol, file, or concept in the codebase",
-        "docs understand symbol function class architecture why",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/index",
-        "Build and search a local knowledge-base index of the codebase",
-        "search knowledge base symbol tfidf find query",
-        "general",
-        available=True,
-    ),
-    SlashCommandSpec(
-        "/ambient",
-        "Manage the ambient agent daemon — autonomous scheduled and triggered agents",
-        "daemon service schedule cron background autonomous trigger webhook",
-        "agent",
-        available=True,
-    ),
-)
 
 FEATURED_HELP_COMMANDS_LEFT: tuple[str, ...] = (
     "/help",
