+++ libs/cli/bog_agents_cli/commands/quality.py
@@ -1,0 +1,49 @@
+"""Code quality and review commands."""
+
+from __future__ import annotations
+
+from bog_agents_cli._spec import SlashCommandSpec
+from bog_agents_cli.commands._base import SlashCommand
+
+COMMANDS: tuple[SlashCommand, ...] = (
+    SlashCommand(
+        spec=SlashCommandSpec(
+            "/audit",
+            "Audit dependencies for vulnerabilities",
+            "security deps",
+            "quality",
+            available=True,
+        ),
+        handler_method="_handle_audit_command",
+    ),
+    SlashCommand(
+        spec=SlashCommandSpec(
+            "/recommend",
+            "Run AI-powered code review and recommendation flows",
+            "review audit advise persona focus",
+            "quality",
+            available=True,
+        ),
+        handler_method="_dispatch_recommend_command",
+    ),
+    SlashCommand(
+        spec=SlashCommandSpec(
+            "/review",
+            "Ask the agent for a structured code review",
+            "lint check staged files commit",
+            "quality",
+            available=True,
+        ),
+        handler_method="_handle_review_command",
+    ),
+    SlashCommand(
+        spec=SlashCommandSpec(
+            "/test",
+            "Run tests with coverage and generate test skeletons",
+            "coverage pytest",
+            "quality",
+            available=True,
+        ),
+        handler_method="_handle_test_command",
+    ),
+)
