+++ libs/cli/bog_agents_cli/commands/ui.py
@@ -1,0 +1,29 @@
+"""UI / display preference commands."""
+
+from __future__ import annotations
+
+from bog_agents_cli._spec import SlashCommandSpec
+from bog_agents_cli.commands._base import SlashCommand
+
+COMMANDS: tuple[SlashCommand, ...] = (
+    SlashCommand(
+        spec=SlashCommandSpec(
+            "/silent",
+            "Quiet mode: tool calls show as one-liners (full details still in log)",
+            "quiet collapse mute hide",
+            "ui",
+            available=True,
+        ),
+        handler_method="_handle_silent_command",
+    ),
+    SlashCommand(
+        spec=SlashCommandSpec(
+            "/verbose",
+            "Verbose mode: tool calls show as expandable widgets (default)",
+            "loud expand show",
+            "ui",
+            available=True,
+        ),
+        handler_method="_handle_verbose_command",
+    ),
+)
